fix: lint error
